Reminder dedup, auto-complete cron, REAGEN_NOW link, SQL timezone, fee INT, fuente fix
- Re-add reminder dedup per appointment_id (48h window), bypass with force=1
- Auto-complete cron: marks appointments Completada ~1h after end, creates pending payment
- REAGEN_NOW WhatsApp reply: sends "Nombre, vamos a reprogramar tu cita" + link with phone
- SET time_zone=-04:00 on MySQL session so NOW()/CURDATE() return Bolivia time
- Fee/amount columns changed from DECIMAL to INT (Bolivianos are whole numbers)
- Source/fuente options: only 3 (Referencia de amigos, Redes sociales, Otro) + migration
- Reschedule banner: shows client name at 18px, supports ?r=1 universal mode

+++ server/routes/booking.js
@@ -57,8 +57,8 @@ router.post('/book', bookingLimiter, validate(bookingSchema), async (req, res) =
     const [clients] = await pool.query('SELECT * FROM clients WHERE id = ? AND tenant_id = ?', [check.client_id, tenantId]);
     const client = clients[0];
     // Apply fee override if provided (from ?fee= URL param)
-    if (fee_override && parseFloat(fee_override) > 0) {
-      const newFee = parseFloat(fee_override);
+    if (fee_override && parseInt(fee_override) > 0) {
+      const newFee = parseInt(fee_override);
       await pool.query('UPDATE clients SET fee = ? WHERE id = ?', [newFee, client.id]);
       client.fee = newFee;
       console.log(`[booking] Fee override: client ${client.id} → Bs ${newFee}`);
